azure_functions/requirements.txt
# Azure Functions Python Requirements
# Includes shared dependencies from src/ module

# Azure Functions runtime
azure-functions==1.18.0

# Shared dependencies (same as container)
azure-storage-blob==12.19.0
azure-identity==1.15.0
yfinance==0.2.36
pandas==2.2.0
pyarrow==15.0.0
pyyaml==6.0.1
requests==2.31.0
tenacity==8.2.3

# Additional for Excel support (HKEX data)
openpyxl==3.1.2
lxml==5.1.0
